Prototype maze algorithm
Based on a prim's based algorithm from wiki.

+++ src/u_maze.h
@@ -7,6 +7,6 @@ UMazeData u_maze_gen_grid(unsigned width, unsigned height);
 
 void u_maze_delete_grid(UMazeData *pMazeData);
 
-UMazeConnection* u_maze_gen(UMazeData *pMazeData, unsigned *pEdgeAmount, uint32_t seed);
+UMazeConnection* u_maze_gen(UMazeData *pMazeData, size_t *pEdgeAmount, uint32_t seed);
 
 #endif // U_MAZE_29
